add customized profile page & secure article deletion

+++ delete-article.php
@@ -1,11 +1,22 @@
 <?php
 require_once __DIR__ . "/database/database.php";
-$articleDB = require_once __DIR__ . "/database/models/ArticleDB.php";
+require_once __DIR__ . "/database/security.php";
 
-$_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-$articleId = $_GET["id"] ?? "";
+$currentUser = isLoggedIn();
 
-if ($articleId) {
-    $articleDB->deleteOne($articleId);
+if ($currentUser) {
+    $articleDB = require_once __DIR__ . "/database/models/ArticleDB.php";
+
+    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $articleId = $_GET["id"] ?? "";
+
+    if ($articleId) {
+        $article = $articleDB->fetchOne($articleId);
+        if ($article["article_author"] == $currentUser["user_id"]) {
+            $articleDB->deleteOne($articleId);
+        }
+    }
 }
+
+
 header("Location: /");
